refactor: 컨텍스트 구조 변경 및 관련 훅 리팩토링
