+++ src/contexts/notification/useNotification.ts
@@ -1,0 +1,12 @@
+import { useContext } from "react";
+import { NotificationContext } from "./NotificationContext";
+
+export const useNotification = () => {
+  const context = useContext(NotificationContext);
+  if (context === undefined) {
+    throw new Error(
+      "useNotification must be used within a NotificationProvider"
+    );
+  }
+  return context;
+};
