+++ src/contexts/theme/ThemeContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from "react";
+
+interface ThemeContextType {
+  theme: "light" | "dark";
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined
+);
