+++ src/contexts/user/UserProvider.tsx
@@ -1,20 +1,7 @@
-import React, { createContext, useContext, useState } from "react";
-import { useMemo, useCallback } from "../@lib/hooks";
-import { useNotification } from "./NotificationContext";
-
-interface User {
-  id: number;
-  name: string;
-  email: string;
-}
-
-interface UserContextType {
-  user: User | null;
-  login: (email: string) => void;
-  logout: () => void;
-}
-
-const UserContext = createContext<UserContextType | undefined>(undefined);
+import { useState } from "react";
+import { useMemo, useCallback } from "../../@lib/hooks";
+import { useNotification } from "../notification/useNotification";
+import { UserContext, User } from "./UserContext";
 
 export const UserProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
@@ -49,12 +36,4 @@ export const UserProvider: React.FC<{ children: React.ReactNode }> = ({
   );
 
   return <UserContext.Provider value={value}>{children}</UserContext.Provider>;
-};
-
-export const useUser = () => {
-  const context = useContext(UserContext);
-  if (context === undefined) {
-    throw new Error("useUser must be used within a UserProvider");
-  }
-  return context;
 };
