+++ src/contexts/user/useUser.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { UserContext } from "./UserContext";
+
+export const useUser = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useUser must be used within a UserProvider");
+  }
+  return context;
+};
